setCognitoAttribute is now an env variable
