+++ APILambda/app.ts
@@ -11,7 +11,7 @@ import { joinEventRouter } from "./events/joinEventById.js"
 import { leaveEventRouter } from "./events/leaveEvent.js"
 import { autocompleteUsersRouter } from "./user/autocompleteUsers.js"
 import { createBlockUserRouter } from "./user/blockUser.js"
-import { createUserProfileRouter } from "./user/createUserProfile.js"
+import { createUserProfileRouter } from "./user/createUser/createUserProfile.js"
 import { deleteUserAccountRouter } from "./user/deleteUserAccount.js"
 import { getSelfRouter } from "./user/getSelf.js"
 import { getUserRouter } from "./user/getUser.js"
