+++ APILambda/index.ts
@@ -6,6 +6,7 @@ import { addBenchmarking, addRoutes, createApp } from "./app.js"
 import { addCognitoTokenVerification } from "./auth.js"
 import { ServerEnvironment } from "./env.js"
 import { addErrorReporting } from "./errorReporting.js"
+import { setProfileCreatedAttribute } from "./user/createUser/createUserProfileSetCognitoAttribute.js"
 
 export const addEventToRequest = (app: Express) => {
   app.use((req, res, next) => {
@@ -34,7 +35,12 @@ export const addEventToRequest = (app: Express) => {
   app.use(express.urlencoded({ extended: true }))
 }
 
-const env: ServerEnvironment = { environment: "prod", eventStartWindowInHours: 1, maxArrivals: 100 }
+const env: ServerEnvironment = {
+  environment: "prod",
+  eventStartWindowInHours: 1,
+  maxArrivals: 100,
+  setProfileCreatedAttribute: (userId: string) => setProfileCreatedAttribute(userId)
+}
 
 const app = createApp()
 addEventToRequest(app)
